trip_planner/requirements.txt
# requirements.txt
streamlit==1.31.0
crewai==0.28.0
langchain>=0.1.10,<0.2.0
langchain-community>=0.0.13
langchain-core>=0.1.9,<0.2.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai>=0.3.1,<0.4.0
langchain-google-genai
